Update ProbeService.java readiness to report ready only if the application is running
In kubernetes, readiness are used to determine if a loadbalancer can forward request to a pod.
If a pod declare that he is ready, he will receive traffic.

The readiness for a kafka stream application is useful if we want to deploy a rest endpoint on top of it like an interactive query. (https://docs.confluent.io/platform/current/streams/developer-guide/interactive-queries.html)
In such a case, we do not want a pod to pretend that it is ready if it is rebalancing, since it is potentially transferring state stores and therefore a potentially inconsistent state.